Add files via upload
I have moved the punctuation so as to eliminate the extra space, added a few more <orig> markup and eliminated a few mistakes. I am not sure what I need to do about the fact that the original page numbers show up on the website and also that the notes I used to markup the changes in punctuations also can be seen on the website.